pyproject.toml
[tool.poetry]
name = "faust-test"
description = "A test for faust"
version = "0"
authors = ["Thomas Chauvet"]

[tool.poetry.dependencies]
faust-streaming = "^0.8.6"
python = "~3.9"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
flake8 = "^4.0.1"
isort = { version = "^5.10.1", extras = ["colors"] }
pytest = "^7.1.2"
pytest-asyncio = "^0.19.0"
pytest-cov = "^3.0.0"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
addopts = "-ra -q"
testpaths = ["tests"]
markers = ["slow: marks tests as slow (deselect with '-m \"not slow\"')"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
